Fix GtkListBoxRow 3.14 versioning
