+++ gtk/layout_since_3_14.go
@@ -1,0 +1,50 @@
+// Copyright (c) 2013-2014 Conformal Systems <info@conformal.com>
+//
+// This file originated from: http://opensource.conformal.com/
+//
+// Permission to use, copy, modify, and distribute this software for any
+// purpose with or without fee is hereby granted, provided that the above
+// copyright notice and this permission notice appear in all copies.
+//
+// THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
+// WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF
+// MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
+// ANY SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
+// WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN
+// ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
+// OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+
+// +build !gtk_3_6,!gtk_3_8,!gtk_3_10,!gtk_3_12
+// not use this: go build -tags gtk_3_8'. Otherwise, if no build tags are used, GTK 3.10
+
+package gtk
+
+// #include <gtk/gtk.h>
+// #include "gtk.go.h"
+import "C"
+
+/*
+ * GtkListBoxRow
+ */
+
+// SetActivatable is a wrapper around gtk_list_box_row_set_activatable().
+func (v *ListBoxRow) SetActivatable(activatable bool) {
+	C.gtk_list_box_row_set_activatable(v.native(), gbool(activatable))
+}
+
+// GetActivatable is a wrapper around gtk_list_box_row_get_activatable().
+func (v *ListBoxRow) GetActivatable() bool {
+	c := C.gtk_list_box_row_get_activatable(v.native())
+	return gobool(c)
+}
+
+// SetSelectable is a wrapper around gtk_list_box_row_set_selectable().
+func (v *ListBoxRow) SetSelectable(selectable bool) {
+	C.gtk_list_box_row_set_selectable(v.native(), gbool(selectable))
+}
+
+// GetSelectable is a wrapper around gtk_list_box_row_get_selectable().
+func (v *ListBoxRow) GetSelectable() bool {
+	c := C.gtk_list_box_row_get_selectable(v.native())
+	return gobool(c)
+}
